requirements.txt
langchain
langchain-community
langchain-chroma
langchain-cohere
python-dotenv
chromadb
pypdf